- Tightening up new code

+++ src/test/resources/com/acme/exp/UrlToExpressionDebugTest.groovy
@@ -11,6 +11,6 @@ class UrlToExpressionDebugTest extends NCubeGroovyExpression
             return -1
         }
         LibCode libCode = new LibCode()
-        libCode.pow(input.age, 2)
+        libCode.pow(input.age as double, 2d)
     }
 }
